+++ package.json
@@ -11,6 +11,7 @@
     "email": "hello@feathers.cloud",
     "url": "https://feathers.cloud"
   },
+  "type": "module",
   "license": "MIT",
   "funding": {
     "type": "github",
@@ -33,12 +34,14 @@
     "eslint": "eslint \"packages/**/*.ts\" --fix",
     "lint": "npm run prettier && npm run eslint",
     "update-dependencies": "ncu -u && lerna exec -- ncu -u",
-    "test": "npm run lint && c8 npm test --workspaces"
+    "dev": "vitest watch --coverage",
+    "vitest": "vitest run --coverage",
+    "test": "npm run lint && npm run compile --workspaces && npm run vitest"
   },
   "devDependencies": {
     "@typescript-eslint/eslint-plugin": "^6.17.0",
     "@typescript-eslint/parser": "^6.17.0",
-    "c8": "^9.0.0",
+    "@vitest/coverage-v8": "^1.2.2",
     "eslint": "^8.56.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-import": "^2.29.1",
@@ -47,6 +50,7 @@
     "lerna": "^8.0.1",
     "npm-check-updates": "^16.14.12",
     "prettier": "^3.1.1",
-    "typescript": "^5.3.3"
+    "typescript": "^5.3.3",
+    "vitest": "^1.2.2"
   }
 }
