+++ packages/pinion/test/cli.test.ts
@@ -1,24 +1,18 @@
-import { describe, it } from 'node:test'
+import { describe, it } from 'vitest'
 import assert from 'assert'
 import { fileURLToPath } from 'url'
 import { dirname } from 'path'
-import { cli } from '../lib/index.js'
+import { cli } from '../src/index.js'
 
 const __filename = fileURLToPath(import.meta.url)
 const __dirname = dirname(__filename)
 
 describe('@featherscloud/pinion/cli', () => {
   it('runs the CLI with a generator and command line arguments', async () => {
-    const oldCwd = process.cwd()
-
-    process.chdir(__dirname)
-
-    const ctx = await cli(['templates/cli.ts', '--name', 'testing'])
+    const ctx = await cli(['packages/pinion/test/templates/cli.ts', '--name', 'testing'])
 
     assert.ok(ctx.noop)
     assert.strictEqual(ctx.name, 'testing')
-
-    process.chdir(oldCwd)
   })
 
   it('errors without generator file', async () => {
