+++ packages/pinion/test/templates/noop.ts
@@ -1,3 +1,3 @@
-import { PinionContext } from '../../lib/index.js'
+import { PinionContext } from '../../src/index.js'
 
 export const generate = (ctx: PinionContext) => Promise.resolve(ctx)
