+++ vitest.config.ts
@@ -1,0 +1,11 @@
+// vitest.config.ts
+import { defineConfig } from 'vitest/config'
+
+export default defineConfig({
+  test: {
+    coverage: {
+      provider: 'v8',
+      include: ['packages/*/src/**/*.ts']
+    }
+  }
+})
