+++ vitest.workspace.ts
@@ -1,0 +1,4 @@
+import { defineWorkspace } from 'vitest/config'
+
+// defineWorkspace provides a nice type hinting DX
+export default defineWorkspace(['packages/*'])
